backend/requirements.txt
aiosqlite==0.20.0
asyncpg==0.30.0
bcrypt==4.3.0
email-validator==2.2.0
fastapi==0.115.6
google-generativeai==0.8.3
httpx==0.28.1
pydantic-settings==2.7.1
pypdf==5.1.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
SQLAlchemy==2.0.36
uvicorn[standard]==0.34.0
